# packages/api-contracts/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenMake.Ai API",
    "description": "\nAI 채팅 어시스턴트 API 문서\n\n## 기능\n- 채팅 메시지 전송 및 스트리밍 응답\n- 파일 업로드 및 문서 분석\n- 클러스터 관리\n- 사용자 인증\n\n## 인증\n대부분의 API는 JWT 토큰 인증이 필요합니다.\n`Authorization: Bearer <token>` 헤더를 사용하세요.\n\n## CSRF (비브라우저 클라이언트)\nBearer 헤더 요청은 CSRF 검증에서 면제됩니다. 단 **사전 인증 mutating 요청**\n(login/register/refresh)은 `GET /api/csrf-token` 으로 토큰을 발급받아\n발급 쿠키와 `X-CSRF-Token` 헤더를 함께 보내야 합니다 (Double-Submit).\n",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "사용 가능 모델 목록",
        "description": "로컬 모델 카탈로그(기본 모델이 첫 entry) + 인증 사용자의 외부 provider 등록 모델을 반환합니다. 비인증도 로컬 카탈로그는 조회 가능.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "parameters": [
          {
            "name": "usableOnly",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "description": "1 이면 역할 배정 가능(streaming+toolCalling) 모델만 반환"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "defaultModel",
                        "models"
                      ],
                      "properties": {
                        "defaultModel": {
                          "type": "string",
                          "description": "기본 chat 모델 full id"
                        },
                        "models": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ModelEntry"
                          }
                        },
                        "imageModel": {
                          "type": "string",
                          "nullable": true,
                          "description": "이미지 생성 모델 id (미설정 시 null)"
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "모델 카탈로그 조회 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/model-assignments": {
      "get": {
        "tags": [
          "Model Assignments"
        ],
        "summary": "내 모델 배정 목록",
        "description": "슬롯 목록 + 본인 배정 + 실효(effective) 해석을 반환합니다.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ModelAssignmentsResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/model-assignments/{slot}": {
      "put": {
        "tags": [
          "Model Assignments"
        ],
        "summary": "슬롯 배정",
        "description": "text 슬롯은 역할 규칙(BYOK·채팅 가능 모델), modality 슬롯은 기능 규칙으로 검증합니다.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slot",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "슬롯 id (config/model-slots)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelAssignmentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "assignment"
                      ],
                      "properties": {
                        "assignment": {
                          "$ref": "#/components/schemas/ModelSlotAssignment"
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "검증 실패(슬롯·모델·키)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Model Assignments"
        ],
        "summary": "슬롯 배정 해제",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slot",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "슬롯 id (config/model-slots)"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "deleted"
                      ],
                      "properties": {
                        "deleted": {
                          "type": "boolean"
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "배정 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/model-assignments": {
      "get": {
        "tags": [
          "Model Assignments"
        ],
        "summary": "전역 모델 배정 목록 (admin)",
        "description": "scope=__global__. 슬롯 목록 + 전역 배정 + 전역 티어 실효 해석.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ModelAssignmentsResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/model-assignments/{slot}": {
      "put": {
        "tags": [
          "Model Assignments"
        ],
        "summary": "전역 슬롯 배정 (admin)",
        "description": "전역은 서버 공용 키 규칙으로 검증. 비배정 슬롯(chat·router)도 허용.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slot",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "슬롯 id (config/model-slots)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelAssignmentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "assignment"
                      ],
                      "properties": {
                        "assignment": {
                          "$ref": "#/components/schemas/ModelSlotAssignment"
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "검증 실패(슬롯·모델·키)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Model Assignments"
        ],
        "summary": "전역 슬롯 배정 해제 (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slot",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "슬롯 id (config/model-slots)"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "deleted"
                      ],
                      "properties": {
                        "deleted": {
                          "type": "boolean"
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ApiMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "배정 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiMeta": {
        "type": "object",
        "required": [
          "timestamp"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "ApiFailure": {
        "type": "object",
        "required": [
          "success",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "ModelCapabilities": {
        "type": "object",
        "required": [
          "executionStrategy",
          "thinking",
          "discussion",
          "vision",
          "toolCalling",
          "streaming"
        ],
        "properties": {
          "executionStrategy": {
            "type": "string",
            "enum": [
              "single"
            ]
          },
          "thinking": {
            "type": "string",
            "enum": [
              "off",
              "medium"
            ]
          },
          "discussion": {
            "type": "boolean"
          },
          "vision": {
            "type": "boolean"
          },
          "toolCalling": {
            "type": "boolean"
          },
          "streaming": {
            "type": "boolean"
          }
        }
      },
      "ModelEntry": {
        "type": "object",
        "required": [
          "name",
          "modelId",
          "description",
          "provider",
          "capabilities"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "modelId": {
            "type": "string",
            "description": "provider prefix 포함 full model id"
          },
          "description": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "capabilities": {
            "$ref": "#/components/schemas/ModelCapabilities"
          },
          "available": {
            "type": "boolean",
            "description": "startup probe 결과 — false 면 UI dimmed"
          },
          "unavailableReason": {
            "type": "string"
          },
          "isFree": {
            "type": "boolean"
          },
          "pricing": {
            "type": "object",
            "properties": {
              "input": {
                "type": "number"
              },
              "output": {
                "type": "number"
              }
            }
          }
        }
      },
...
      "ModelSlotInfo": {
        "type": "object",
        "required": [
          "id",
          "group",
          "kind",
          "roles",
          "capabilities",
          "paramKeys",
          "available"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "group": {
            "type": "string",
            "enum": [
              "agents",
              "quality",
              "multimodal"
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "text",
              "modality"
            ],
            "description": "text=채팅 가능 모델 · modality=전체 모델"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "이 슬롯을 읽는 역할"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "이 슬롯을 읽는 기능"
          },
          "paramKeys": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "저장 가능한 params 키(없으면 빈 배열)"
          },
          "available": {
            "type": "boolean",
            "description": "실행 가능 여부 — 어댑터 없음/소유 add-on 꺼짐이면 false"
          }
        }
      },
      "ModelSlotAssignment": {
        "type": "object",
        "required": [
          "slot",
          "fullId",
          "params",
          "updatedAt"
        ],
        "properties": {
          "slot": {
            "type": "string"
          },
          "fullId": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ModelSlotEffective": {
        "type": "object",
        "required": [
          "slot",
          "fullId",
          "source"
        ],
        "properties": {
          "slot": {
            "type": "string"
          },
          "fullId": {
            "type": "string",
            "nullable": true,
            "description": "지금 실제로 쓰이는 모델 — 해석 실패면 null"
          },
          "source": {
            "type": "string",
            "enum": [
              "user",
              "global",
              "default",
              "none"
            ]
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "ModelAssignmentsResponse": {
        "type": "object",
        "required": [
          "slots",
          "assignments",
          "effective"
        ],
        "properties": {
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelSlotInfo"
            }
          },
          "assignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelSlotAssignment"
            }
          },
          "effective": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelSlotEffective"
            }
          }
        }
      },
      "ModelAssignmentInput": {
        "type": "object",
        "required": [
          "model"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "provider:model 또는 로컬 태그"
          },
          "params": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key (omk_live_...) — X-API-Key 헤더 또는 Authorization: Bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ]
}